build: switch to Terraform Cloud app on GitHub
Switch the current setup of using GitHub Actions for Terraform "plan"
and "apply" runs to using the official Terraform Cloud app for GitHub.

Replace the old starter action with a "Terraform format check" that only
checks if `terraform fmt` was run locally.

Fix formatting issues in `team-members.tf`.

The "Terraform format check" check as well as the "Terraform format
check" checks are added to the "terraform-main"
`github_branch_protection`.